refactored some files to be more atomic and precise in their function, serialize row is on its own, tf_lstm main refactored to functions, todo move to proper helper files for general usage
