+++ sips/lines/bov/better/bet_utils.py
@@ -277,6 +277,79 @@ def get_team_names(driver):
     return team_names
 
 
+def login(driver=None, captcha_wait=60):
+    '''
+    goes to bovadas login and gets you to captcha, then waits for a time
+    '''
+    if not driver:
+        driver = base_driver()
+
+    driver.get(ROOT_URL + '?overlay=login')
+    time.sleep(0.5)
+
+    inputs = driver.find_elements_by_tag_name('input')
+
+    email_field = inputs[0]
+    email_field.send_keys('dahmeh.almos@gmail.com')
+
+    password_field = inputs[1]
+    password_field.send_keys('BeautifulBeachesinLagos5')
+
+    login_field = inputs[2]
+    login_field.send_keys('\n')
+
+    print(f'waiting for {captcha_wait} seconds for user to fill out captcha')
+    time.sleep(captcha_wait)
+    # driver.close()
+
+
+def get_driver(sport='football/nfl', sleep=1.5):
+    '''
+    opens chrome to nfl lines
+    '''
+
+    driver = base_driver()
+    driver.get(ROOT_URL + 'sports/' + sport)
+
+    if sleep:
+        time.sleep(sleep)
+
+    return driver
+
+
+def bet(bet_tup, index, driver=None, verbose=False):
+    '''
+    bet_tup consists of team_name, mkt_type, amt
+
+    ideally input args are (driver, game_id, amt)
+    and returns the index of the top-line where it is currently located
+    on the bet slip
+    '''
+    team_name, mkt_type, amt = bet_tup
+    button = find_bet_button(team_name, mkt_type, driver, verbose)
+    button.send_keys("\n")
+    time.sleep(1)
+    set_wager(driver, index, amt)
+    time.sleep(1)
+
+
+def delete_bets(driver, indices):
+    '''
+    deletes the bets from the betslip based on the top-line indices you specify
+    '''
+    if isinstance(indices, int):
+        indices = [indices]
+
+    top_lines = driver.find_elements_by_class_name(u.TOP_LINE_CLASS)
+
+    for i in indices:
+        top_line = top_lines[i]
+        delete_bet_button = top_line.find_element_by_tag_name('button')
+        delete_bet_button.send_keys('\n')
+        time.sleep(1)
+        print(f'bet[{i}] deleted')
+
+
 def base_driver(screen=None):
     '''
     screen is either None, 'full', or 'min'
